Enhance inventory management in Order and Product classes
- Added `CompletionInventoryApplied` property to track inventory adjustments on order completion.
- Changed `StockQuantity` type to `decimal?` in `Product` and `ProductVariant` for fractional support.
- Introduced `VariationStockByQuantity` in `Product` to manage stock per variation.
- Updated service methods in `OrderService` and `ProductService` for new properties.
- Added migration scripts for database schema updates.
- Adjusted WooCommerce integration for improved stock management.

+++ Core/George.Data/Dto/ProductDto.cs
@@ -12,7 +12,7 @@ namespace George.Data.Dto
         public Dictionary<string, string>? OptionValues { get; set; }
         public decimal? Price { get; set; }
         public decimal? SalePrice { get; set; }
-        public int? StockQuantity { get; set; }
+        public decimal? StockQuantity { get; set; }
         public string? Sku { get; set; }
         public decimal? Weight { get; set; }
     }
